somgrid {class}	R Documentation

Plot SOM Fits

Description

Plotting functions for SOM results.

Usage

somgrid(xdim = 8, ydim = 6, topo = c("rectangular", "hexagonal"))

## S3 method for class 'somgrid':
plot(x, type = "p", ...)

## S3 method for class 'SOM':
plot(x, ...)

Arguments

xdim, ydim 	 dimensions of the grid

topo 	 the topology of the grid.

x 	 an object inheriting from class "somgrid" or "SOM".

type, ... 	 graphical parameters.

Details

The class "somgrid" records the coordinates of the grid to be used for (batch or on-line) SOM: this has a plot method.

The plot method for class "SOM" plots a stars plot of the representative at each grid point.

Value

For somgrid, an object of class "somgrid", a list with components

pts 	 a two-column matrix giving locations for the grid points.

xdim, ydim, topo 	 as in the arguments to somgrid.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

batchSOM, SOM

[Package class version 7.2-48 Index]